HawAI-backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
motor==3.6.0
passlib[bcrypt]==1.7.4
PyJWT==2.9.0
httpx==0.27.2
python-multipart==0.0.9
pydantic-settings==2.5.2
pydantic==2.9.2
pytest==8.3.3
pytest-asyncio==0.24.0
email-validator==2.2.0
